Substantial rewrite of restriction checking system and item + mount parsing
Restrictions are now unified across the whole codebase - using a common interface
of std::unordered_map<RestrictionField, std::any> wherever it is needed. Here the
std::any value type is initialised and cast according the the value of the
RestrictionField enum instance stored as the key type.

Parsing of items and mounts is being changed over to the new format already
being used for roster parsing. Storage of mount options and restrictions is
now handled by mount_option rather than the mount class - the latter now just
storing the base properties of a mount (i.e. characteristics, special rules).

NOTE: this commit is a WIP, building of the project is guaranteed to fail on this commit.